add certificate pdf file
